requirements.txt
streamlit
langchain
openai
PyPDF2
python-dotenv
faiss-cpu
huggingface_hub
sentence-transformers